Performance improvements, small cleanup
